% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics-utils.R
\name{cohen_d}
\alias{cohen_d}
\title{Cohen's d and Hedges g effect size}
\usage{
cohen_d(...)
}
\arguments{
\item{...}{A comma separated list of arguments. See \code{\link[effsize:cohen.d]{effsize::cohen.d()}}.}
}
\value{
A list of class \code{effsize} as returned by \code{\link[effsize:cohen.d]{effsize::cohen.d()}}.
}
\description{
This is wrapper to the \code{\link[effsize:cohen.d]{effsize::cohen.d()}} function.
}
\examples{
cohen_d(weight ~ gender, data = ansur)
cohen_d(age ~ gender, data = schizophrenia)
}
