% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics-utils.R
\name{cor_test}
\alias{cor_test}
\title{Test for Correlation Between Paired Samples}
\usage{
cor_test(x, y, method = "pearson", data)
}
\arguments{
\item{x}{A numeric variable.}

\item{y}{A numeric variable.}

\item{method}{A character string indicating which correlation
coefficient is to be used: "pearson", "kendall", or "spearman". Default method is "pearson".}

\item{data}{A data frame containing the \code{y} and \code{x} variables}
}
\value{
A tibble data frame with the correlation statistic, and the corresponding p-value.
}
\description{
This function is a wrapper around \code{\link[stats:cor.test]{stats::cor.test()}}.
It implements the Pearson's correlation test that tests the null hypothesis
that two paired samples of values are unrelated.
This function must be applied to two numeric vectors.
}
\examples{
cor_test(y = sex_dimorph, x = attractive, data = faithfulfaces)
cor_test(y = sex_dimorph, x = attractive, method = "spearman", data = faithfulfaces)
}
