% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling-utils.R
\name{drop_if_all_na}
\alias{drop_if_all_na}
\title{Drop rows if all values on selected columns are missing}
\usage{
drop_if_all_na(data, ...)
}
\arguments{
\item{data}{A data frame}

\item{...}{<\code{\link[=tidyr_tidy_select]{tidy-select}}> Columns to inspect for missing values. If empty, all columns are used.}
}
\value{
A data frame, possibly with some rows dropped.
}
\description{
Remove a row if all values on selected columns, or by default, on all
columns, are missing, i.e. have values of NA or NaN.
}
\details{
The \link[tidyr]{drop_na} function will remove any row if it has any NA in selected columns.
By default, it will remove the row there is any NA or NaN in any column.
This \code{drop_if_all_na} function is similar but removes the row only if all values in the selected columns are NA or NaN.
As with \link[tidyr]{drop_na}, by default it will use all columns.
In other words, by default, \code{drop_if_all_na} removes any row if all values on that row are NA or NaN.
}
\examples{
data_df <- data.frame(x = c(1, 2, NA, NA), y = c(2, NA, 5, NA))

drop_if_all_na(data_df)
drop_if_all_na(data_df, x)
drop_if_all_na(data_df, y)
drop_if_all_na(data_df, x, y)
drop_if_all_na(data_df, x:y)
drop_if_all_na(data_df, starts_with('x'), ends_with('y'))

}
