% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{faithfulfaces}
\alias{faithfulfaces}
\title{Faithfulness from a Photo?}
\format{
A data frame with 170 observations on the following 7 variables.
\describe{
\item{sex_dimorph}{Rating of sexual dimorphism (masculinity for males, femininity for females)}
\item{attractive}{Rating of attractiveness}
\item{cheater}{Was the face subject unfaithful to a partner?}
\item{trustworthy}{Rating of trustworthiness}
\item{faithful}{Rating of faithfulness}
\item{face_sex}{Sex of face (female or male)}
\item{rater_sex}{Sex of rater (female or male)}
}
}
\source{
This data set was taken from the
\href{https://cran.r-project.org/package=Stat2Data}{\code{Stats2Data}
R package}. From the description in that package, the original is based on
G. Rhodes et al. (2012), "Women can judge sexual unfaithfulness from
unfamiliar men's faces," Biology Letters, November 2012. All of the 68
raters were heterosexual Caucasians, as were the 170 persons who were
rated. (We have deleted 3 subjects with missing values and 16 subjects who
were over age 35.)
}
\usage{
faithfulfaces
}
\description{
Ratings from a facial photo and actual faithfulness.
}
\details{
College students were asked to look at a photograph of an opposite-sex adult
face and to rate the person, on a scale from 1 (low) to 10 (high), for
attractiveness. They were also asked to rate trustworthiness, faithfulness,
and sexual dimorphism (i.e., how masculine a male face is and how feminine a
female face is). Overall, 68 students (34 males and 34 females) rated 170
faces (88 men and 82 women).
}
\keyword{datasets}
