% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics-utils.R
\name{get_dummy_code}
\alias{get_dummy_code}
\title{Show the dummy code of a categorical variable}
\usage{
get_dummy_code(Df, variable)
}
\arguments{
\item{Df}{A data frame}

\item{variable}{A categorical variable (e.g. character vector or factor).}
}
\value{
A data frame whose rows provide the dummy code for
each distinct value of \code{variable}.
}
\description{
For each value of a categorical variables, show the binary
code used in a regression model to represent its value.
This is wrapper to the \code{\link[fastDummies:dummy_cols]{fastDummies::dummy_cols()}} function.
}
\examples{
get_dummy_code(PlantGrowth, group)
}
