% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-utils.R
\name{histogram}
\alias{histogram}
\title{A histogram}
\usage{
histogram(x, data, by = NULL, position = "stack", bins = 10, alpha = 1)
}
\arguments{
\item{x}{The numeric variable that is to be histogrammed.}

\item{data}{A data frame with at least one numeric variable (the \code{x}
variable).}

\item{by}{A categorical variable by which to group the \code{x} values. If
provided there will be one histogram for each set of \code{x} values grouped by
the values of the \code{by} variable.}

\item{position}{If the \code{by} variable is provided, there are three ways these multiple
histograms can be positioned: stacked (\code{position = 'stack'}), side by side
(\code{position = 'dodge'}), superimposed (\verb{position = identity'}).}

\item{bins}{The number of bins to use in the histogram.}

\item{alpha}{The transparency to for the filled histogram bars. This is probably
only required when using \code{position = 'identity'}.}
}
\description{
This is a wrapper to the typical \code{ggplot} based histogram, i.e., using
\code{geom_histogram}. A continuous variable, \code{x}, is required as an input.
Optionally, a \code{by} categorical variable can be provided.
}
\examples{
histogram(schizophrenia, x= age, by = gender, bins = 20)
histogram(schizophrenia, x= age, by = gender, position = 'identity', bins = 20, alpha = 0.7)
histogram(schizophrenia, x= age, by = gender, position = 'dodge', bins = 20)
}
