% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jobsatisfaction}
\alias{jobsatisfaction}
\title{Job Satisfaction Data for Two-Way ANOVA}
\format{
A data frame with 58 rows and 3 columns.
}
\usage{
data("jobsatisfaction")
}
\description{
Contains the job satisfaction score organized by gender and education level.
This data set was taken from the
\href{https://cran.r-project.org/package=datarium}{\code{datarium} R
package}.
}
\examples{
data(jobsatisfaction)
jobsatisfaction
}
