% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pairedsleep}
\alias{pairedsleep}
\title{Paired sleep data}
\format{
A data frame with 10 observations on the following 3 variables.
\describe{
\item{ID}{The patient ID.}
\item{y1}{The increase in hours, relative to control, for drug 1.}
\item{y2}{The increase in hours, relative to control, for drug 2.}
}
}
\source{
This data is a transformed version of \link[datasets:sleep]{datasets::sleep}.
}
\usage{
pairedsleep
}
\description{
Data which show the effect of two soporific drugs (increase in hours of
sleep compared to control) on 10 patients.
}
\keyword{datasets}
