% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychometrics.R
\name{re_code}
\alias{re_code}
\title{Recode specified values by new values}
\usage{
re_code(x, from, to)
}
\arguments{
\item{x}{A vector, including column of data frame}

\item{from}{The set of old values to be replaced by new ones}

\item{to}{The set of new values to replace the old ones}
}
\value{
A vector that is the original vector but with old values replaced by new ones
}
\description{
Recode specified values by new values
}
\examples{
# Replace any occurrence of 1 and 2 with 101 and 201, respectively
x <- c(1, 2, 3, 4, 5, 1, 2)
re_code(x, from = c(1, 2), to = c(101, 201))
}
