% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schizophrenia}
\alias{schizophrenia}
\title{Age of Onset of Schizophrenia Data}
\format{
A data frame with 251 observations on the following 2 variables.
\describe{
\item{age}{Age at the time of diagnosis.}
\item{gender}{A categorical variable with values \code{female} and \code{male}}
}
}
\source{
This data set was taken from the
\href{https://cran.r-project.org/package=HSAUR}{\code{HSAUR} R
package}. From the description in that package, the original is E.
Kraepelin (1919), \emph{Dementia Praecox and Paraphrenia}.  Livingstone,
Edinburgh.
}
\usage{
schizophrenia
}
\description{
Data on sex differences in the age of onset of schizophrenia.
}
\details{
A sex difference in the age of onset of schizophrenia was noted by Kraepelin
(1919). Subsequently epidemiological studies of the disorder have
consistently shown an earlier onset in men than in women. One model that has
been suggested to explain this observed difference is known as the subtype
model which postulates two type of schizophrenia, one characterised by early
onset, typical symptoms and poor premorbid competence, and the other by late
onset, atypical symptoms, and good premorbid competence.  The early onset
type is assumed to be largely a disorder of men and the late onset largely a
disorder of women.
}
\keyword{datasets}
