% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{selfesteem}
\alias{selfesteem}
\title{Self-Esteem Score Data for One-way Repeated Measures ANOVA}
\format{
A data frame with 10 rows and 4 columns.
}
\usage{
data("selfesteem")
}
\description{
The dataset contains 10 individuals' self-esteem score on three time points
during a specific diet to determine whether their self-esteem improved.

One-way repeated measures ANOVA can be performed in order to determine the
effect of time on the self-esteem score.

This data set was taken from the
\href{https://cran.r-project.org/package=datarium}{\code{datarium} R
package}.
}
\examples{
data(selfesteem)
selfesteem
}
