% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{selfesteem2}
\alias{selfesteem2}
\title{Self Esteem Score Data for Two-way Repeated Measures ANOVA}
\format{
A data frame with 24 rows and 5 columns.
}
\usage{
data("selfesteem2")
}
\description{
Data are the self esteem score of 12 individuals enrolled in 2
successive short-term trials (4 weeks) - control (placebo) and special diet
trials.

The self esteem score was recorded at three time points: at the beginning
(t1), midway (t2) and at the end (t3) of the trials.

The same 12 participants are enrolled in the two different trials with
enough time between trials.

Two-way repeated measures ANOVA can be performed in order to determine
whether there is interaction between time and treatment on the self
esteem score.

This data set was taken from the
\href{https://cran.r-project.org/package=datarium}{\code{datarium} R
package}.
}
\examples{
data(selfesteem2)
selfesteem2
}
