% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics-utils.R
\name{t_test}
\alias{t_test}
\title{Independent samples t-test}
\usage{
t_test(formula, data)
}
\arguments{
\item{formula}{A two sided formula with one variable on either side, e.g. y ~
x, where the left hand side, dependent, variable is a numeric variable in
\code{data} and the right hand side, independent, variable is a categorical or
factor variable in \code{data}, and which has only two distinct values.}

\item{data}{A data frame that contains the dependent and independent
variables.}
}
\value{
A list with class "htest" as returned by \code{\link[stats:t.test]{stats::t.test()}}.
}
\description{
A wrapper to \code{\link[stats:t.test]{stats::t.test()}} with \code{var.equal = TRUE}.
}
\examples{
t_test(trustworthy ~ face_sex, data = faithfulfaces)
  
}
