% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_psychometrics}
\alias{test_psychometrics}
\title{Psychometrics raw data from testing or demo purposes}
\format{
A data frame with 44 rows and 30 columns
}
\usage{
data('test_psychometrics')
}
\description{
Typical psychometrics raw data files have multiple psychometric
variables (scales), each with multiple constituent items.
In this data set, there are three psychometric variables, each with 10 constituent items.
The variables can be labelled \code{x}, \code{y}, and \code{z}.
The constituent items of \code{x}, \code{y} and \code{z} are \verb{x_1, x_2 ... x_10},
\verb{y_1, y_2 ... y_10}, \verb{z_1, z_2 ... z_10}, respectively.
}
\examples{
data(test_psychometrics)
test_psychometrics
}
\keyword{datasets}
