% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vizverb}
\alias{vizverb}
\title{Visual versus Verbal Perception and Responses}
\format{
A data frame with 80 observations on the following 5 variables.
\describe{
\item{subject}{Subject identifying number (\code{s1} to \code{s20})}
\item{task}{Describe a diagram (\code{visual}) or a sentence (\code{verbal})}
\item{response}{Point response (\code{visual}) or say response (\code{verbal})}
\item{time}{Response time (in seconds)}
}
}
\source{
This data set was taken from the
\href{https://cran.r-project.org/package=Stat2Data}{\code{Stats2Data}
R package}. From the description in that package, the original data appear
to have been collected in a Mount Holyoke College psychology class based
replication of an experiment by Brooks, L., R. (1968) "Spatial and verbal
components of the act of recall," Canadian J. Psych. V 22, pp. 349 - 368.
}
\usage{
vizverb
}
\description{
An experiment studying the interaction between visual versus perception and visual versus verbal responses.
}
\details{
Subjects carried out two kinds of tasks. One task was visual (describing a
diagram), and the other was classed as verbal (reading and describing a
sentence sentences). They reported the results either by pointing (a "visual"
response), or speaking (a verbal response). Time to complete each task was
recorded in seconds.
}
\keyword{datasets}
