% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychometrics.R
\name{cronbach}
\alias{cronbach}
\title{Calculate Cronbach's alpha for sets of psychometric scale items}
\usage{
cronbach(.data, ...)
}
\arguments{
\item{.data}{}

\item{...}{}
}
\value{
A vector of Cronbach alpha scores, with one score per each set of items.
}
\description{
This function calculates the Cronbach alpha for one
or more sets of psychometric scale items. Each item
is a variable in a data frame. Each set of items is
defined by a tidy selection of a set of items.
}
\examples{
 cronbach(test_psychometrics, 
          x = starts_with('x'), 
          y = y_1:y_10) 

}
