% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics-utils.R
\name{ez_anova}
\alias{ez_anova}
\title{Analysis of variance}
\usage{
ez_anova(
  data,
  dv,
  wid,
  within = NULL,
  within_full = NULL,
  within_covariates = NULL,
  between = NULL,
  between_covariates = NULL,
  observed = NULL,
  diff = NULL,
  reverse_diff = FALSE,
  type = 2,
  white.adjust = FALSE,
  detailed = FALSE,
  return_aov = FALSE
)
}
\description{
This is wrapper to the \code{\link[ez:ezANOVA]{ez::ezANOVA()}} function.
}
\examples{
ez_anova(data = selfesteem2_long,
            dv = score,
            wid = id,
            within = c(time, treatment),
            detailed = TRUE,
            return_aov = TRUE)
}
