% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-utils.R
\name{histogram}
\alias{histogram}
\title{A histogram}
\usage{
histogram(
  x,
  data,
  by = NULL,
  position = "stack",
  facet = NULL,
  facet_type = "wrap",
  bins = 10,
  alpha = 1
)
}
\arguments{
\item{x}{The numeric variable that is to be histogrammed.}

\item{data}{A data frame with at least one numeric variable (the \code{x}
variable).}

\item{by}{A categorical variable by which to group the \code{x} values. If
provided there will be one histogram for each set of \code{x} values grouped by
the values of the \code{by} variable.}

\item{position}{If the \code{by} variable is provided, there are three ways these
multiple histograms can be positioned: stacked (\code{position = 'stack'}), side
by side (\code{position = 'dodge'}), superimposed (\verb{position = identity'}).}

\item{facet}{A character string or character vector. If provided, we
\code{facet_wrap} (by default) the histogram by the variables. This is
equivalent to the \code{facet_wrap(variables)} in \code{ggplot2}.}

\item{facet_type}{By default, this takes the value of \code{wrap}, and \code{facet}
leads to a facet wrap. If \code{facet_type} is \code{grid}, then \code{facet} gives us a
\code{facet_grid}.}

\item{bins}{The number of bins to use in the histogram.}

\item{alpha}{The transparency to for the filled histogram bars. This is
probably only required when using \code{position = 'identity'}.}
}
\description{
This is a wrapper to the typical \code{ggplot} based histogram, i.e., using
\code{geom_histogram}. A continuous variable, \code{x}, is required as an input.
Optionally, a \code{by} categorical variable can be provided.
}
\examples{
histogram(x= age, data = schizophrenia, by = gender, bins = 20)
histogram(x= age, data = schizophrenia, by = gender, position = 'identity', bins = 20, alpha = 0.7)
histogram(x= age, data = schizophrenia, by = gender, position = 'dodge', bins = 20)
histogram(x = weight, bins = 20, data = ansur, facet = height_tercile)
histogram(x = weight, bins = 20, data = ansur, 
          facet = c(height_tercile, age_tercile), facet_type = 'grid')
}
